moved logging and added readme
