+++ libs/shared/util-blockchain/tsconfig.json
@@ -14,6 +14,8 @@
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "noImplicitReturns": true,
+    "esModuleInterop": true,
+    "resolveJsonModule": true,
     "noFallthroughCasesInSwitch": true
   }
 }
